+++ .gitignore
@@ -1,7 +1,7 @@
 .parcel-cache
 dist
 node_modules
-idea
+.idea
 coverage
 .env
 .env.development
