+++ src/components/index.js
@@ -1,0 +1,3 @@
+export * from './organisms'
+export * from './atoms'
+export * from './molecules'
