+++ src/pages/Home.js
@@ -1,5 +1,3 @@
-const Home = () => {
+export function Home() {
     return <h1>Home Page</h1>
 }
-
-export default Home
