+++ src/pages/NewGames.js
@@ -1,5 +1,3 @@
-const NewGames = () => {
+export function NewGames() {
     return <h1>New Games Page</h1>
 }
-
-export default NewGames
