+++ src/pages/PopularGames.js
@@ -1,5 +1,3 @@
-const PopularGames = () => {
+export function PopularGames() {
     return <h1>Popular Games Page</h1>
 }
-
-export default PopularGames
