+++ src/pages/UpcomingGames.js
@@ -1,5 +1,3 @@
-const UpcomingGames = () => {
+export function UpcomingGames() {
     return <h1>Upcoming Games Page</h1>
 }
-
-export default UpcomingGames
