+++ src/pages/index.js
@@ -1,0 +1,4 @@
+export { Home } from './Home'
+export { PopularGames } from './PopularGames'
+export { NewGames } from './NewGames'
+export { UpcomingGames } from './UpcomingGames'
